+++ app/web/src/pages/_app.tsx
@@ -32,6 +32,8 @@ import { TagList } from 'src/components/tags/TagList';
 import { ToastContainer } from 'src/components/common/container/ToastContainer';
 import { ApiService } from 'src/services/api.service';
 import { seoConfig } from 'src/utils/seo/config';
+import { loadStyle } from 'src/utils/scripts/load';
+import { isBrowser } from 'src/utils/env';
 
 const theme = {
     colors: {
@@ -102,6 +104,9 @@ class NewsApp extends App<any, any, any> {
         this.setState({
             deviceWidth: 0,
         });
+        loadStyle(
+            'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.9.0/css/all.min.css'
+        );
         Router.events.on('routeChangeComplete', () => {
             window.scrollTo(0, 0);
         });
@@ -235,14 +240,14 @@ class NewsApp extends App<any, any, any> {
     }
 }
 
-if (typeof window === 'undefined') {
+if (typeof window !== 'undefined') {
+    import('src/browser');
+} else {
     import('src/setup').then(({ logger, Sentry }) => {
         process.once('unhandledRejection', (e) => {
             logger.error('Unhandled rejections: ', e);
             Sentry.captureException(e);
         });
     });
-} else {
-    import('src/browser');
 }
 export default withContainer(init)(NewsApp);